feat : 6주차 실습 완료
- pratice-todo 파일에 강의 내용대로 실습완료
