+++ KS-최경수/6주차 미션/pratice-todo/src/App.jsx
@@ -1,4 +1,5 @@
-import { useState } from 'react'
+import { useState, useContext } from 'react'
+import { TodoContext } from './context/TodoContext';
 import './App.css'
 import calenderIcon from './images/calendar_month.png'
 import Header from './components/Header';
@@ -7,42 +8,8 @@ import Items from './components/Items';
 
 
 function App() {
-  // Data
-  const [todos, setTodos] = useState([
-    {id:1, task:"first"},{id:2, task:"second"},{id:3, task:"third"}
-  ]);
-
-  // Header
-  const date = new Date();
-  const dateFormatted = `${date.getFullYear()}.${date.getMonth()+1}.${date.getDate()}`;
-
-  // Component UseState
-  const [text, setText] = useState('');
-  const [editingId, setIsEditingId] = useState('');
-  const [editText, setEditText] = useState('');
-
-  // 1. 추가하기
-  const addTodo = () => {
-    setTodos((prev) => [
-      ...prev,
-      {id : Math.floor(Math.random()*100)+2, task: text},
-    ])
-    setText('');
-  };
-
-  // 2. 삭제하기
-  const deleteTodo = (id) => {
-    setTodos((prev) => prev.filter((item) => item.id !== id));
-  };   
-
-  // 3. 수정하기
-  const updateTodo = () => {
-    console.log(editingId);
-    console.log(editText);
-    setTodos((prev) => prev.map((item) => item.id === editingId ? {...item, task:editText} : item))
-    setIsEditingId('');
-  };
-
+  const {dateFormatted, todos, text, setText, editingId, setIsEditingId, setEditText,
+     addTodo, deleteTodo, updateTodo} = useContext(TodoContext);
 
   return (    
     <>
