+++ KS-최경수/6주차 미션/pratice-todo/src/main.jsx
@@ -2,7 +2,11 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
+import { TodoContextProvider } from './context/TodoContext.jsx'
 
 createRoot(document.getElementById('root')).render(
-    <App />
+    <TodoContextProvider>
+           <App /> 
+    </TodoContextProvider>
+
 )
